refactor(pdf): isolate figure source lineage
